% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneflow.R
\name{as.tuneflow}
\alias{as.tuneflow}
\title{Create a tuneflow: a workflow with instructions for tuning.}
\usage{
as.tuneflow(wf, ...)
}
\arguments{
\item{...}{args to be passed to tune_wf (e.g., v, grid).}
}
\description{
Create a tuneflow: a workflow with instructions for tuning.
}
