% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{crossfit}
\alias{crossfit}
\title{Cross-fit a workflow; a wrapper for \code{fit_resamples}}
\usage{
crossfit(wf, rset)
}
\arguments{
\item{wf}{a tidymodels workflow.}

\item{rset}{a collection (i.e., "set") of resamples.
E.g., for propenisty models, the output of vfold_cv}
}
\description{
Cross-fit a workflow; a wrapper for \code{fit_resamples}
}
