% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{fit.tuned_boost_spec}
\alias{fit.tuned_boost_spec}
\title{Fit a tuned, boosting specification}
\usage{
\method{fit}{tuned_boost_spec}(object, data, ...)
}
\arguments{
\item{object}{A tuned_boost_spec object.}

\item{data}{A data frame containing the data to fit the model on.}

\item{...}{Additional arguments passed to xgboost::xgb.train.}
}
\description{
Fit a tuned, boosting specification
}
