% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneflow.R
\name{fit.tuneflow}
\alias{fit.tuneflow}
\title{Fit a tuneflow object}
\usage{
\method{fit}{tuneflow}(object, data, ...)
}
\arguments{
\item{...}{passed to fit.workflow (e.g., weights)}
}
\description{
Fit a tuneflow object
}
