% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_workflow.R
\name{get_na_workflow}
\alias{get_na_workflow}
\title{A space-filling workflow that always outputs NA predictions.}
\usage{
get_na_workflow()
}
\description{
These are applied as defaults when a user does not supply a workflow.
}
\examples{
wf <- get_na_workflow()
fitted <- fit(wf, mtcars)
preds <- predict(fitted, mtcars)
stopifnot(all(is.na(preds)))
stopifnot(length(preds) == nrow(mtcars))
}
