% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_column.R
\name{get_weights}
\alias{get_weights}
\title{Get the weights vector from a model specification or workflow and data}
\usage{
get_weights(object, data)
}
\arguments{
\item{object}{A model specification or workflow object.}

\item{data}{A data frame containing the data to extract weights from.}
}
\description{
Get the weights vector from a model specification or workflow and data
}
