% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_column.R
\name{get_weights_column}
\alias{get_weights_column}
\title{Get the name of the weights column from a model specification or workflow}
\usage{
get_weights_column(object, ...)
}
\arguments{
\item{object}{A model specification or workflow object.}

\item{...}{Placeholder for generic arguments.}
}
\description{
Get the name of the weights column from a model specification or workflow
}
