% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xy.R
\name{get_x}
\alias{get_x}
\alias{get_y}
\alias{get_x.lightgbm_spec}
\alias{get_x.tuned_boost_spec}
\alias{get_x.cvboost_spec}
\alias{get_y.lightgbm_spec}
\alias{get_y.tuned_boost_spec}
\alias{get_y.cvboost_spec}
\title{Get the design matrix and response vector from a
model specification or workflow and data}
\usage{
get_x(object, data, ...)

get_y(object, data, ...)

\method{get_x}{lightgbm_spec}(object, data, ...)

\method{get_x}{tuned_boost_spec}(object, data, ...)

\method{get_x}{cvboost_spec}(object, data, ...)

\method{get_y}{lightgbm_spec}(object, data, ...)

\method{get_y}{tuned_boost_spec}(object, data, ...)

\method{get_y}{cvboost_spec}(object, data, ...)
}
\arguments{
\item{object}{A model specification or workflow object.}

\item{data}{A data frame containing the data to
extract the design matrix and response vector from.}

\item{...}{Placeholder for generic arguments.}
}
\description{
Get the design matrix and response vector from a
model specification or workflow and data
}
