% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ihdp}
\alias{ihdp}
\title{IHDP Dataset}
\format{
An object of class \code{list} of length 2.
}
\usage{
ihdp
}
\description{
This package contains 100 iterations of the IHDP dataset from Shalit et al. (2017),
which is adapted from Hill (2011). This dataset is copied from https://www.fredjo.com/.
}
\details{
The dataset is a list contains 100 simulations of 672 training observations
and 75 test observations. It contains arrays of treatments (t),
covariates (x), true conditional effects (mu0 and mu1), factual outcomes (yf),
and counterfactual outcomes (ycf).

References:

\itemize{
\item{}{Shalit, U., Johansson, F.D. & Sontag, D.. (2017). Estimating individual treatment effect: generalization bounds and algorithms. \if{html}{\out{<i>}}Proceedings of the 34th International Conference on Machine Learning\if{html}{\out{</i>}}, in \if{html}{\out{<i>}}Proceedings of Machine Learning Research\if{html}{\out{</i>}} 70:3076-3085}
\item{}{Hill, Jennifer L. Bayesian nonparametric modeling for causal inference. Journal of Computational and Graphical Statistics, 20(1), 2011.}
}
}
\keyword{datasets}
