% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{lightgbm_spec}
\alias{lightgbm_spec}
\alias{fit.lightgbm_spec}
\alias{predict.lightgbm_fit}
\title{Define a lightgbm model specification}
\usage{
lightgbm_spec(formula, mode, params = list(), weights_column = NULL)

\method{fit}{lightgbm_spec}(object, data, verbose = -1, ...)

\method{predict}{lightgbm_fit}(object, new_data, ...)
}
\arguments{
\item{formula}{A formula specifying the outcome and predictors for the model.}

\item{mode}{A character string specifying the mode of
the model, either "regression" or "classification".}

\item{params}{A named list of parameters to be passed
to the lightgbm training function.}

\item{weights_column}{An optional string specifying the
name of a column in the training data frame containing observation weights.}

\item{object}{a fitted lightgbm_fit object.}

\item{data}{a data frame containing the training data.}

\item{verbose}{verbosity level for lightgbm training.}

\item{...}{additional arguments passed to the predict method of the lightgbm model.}

\item{new_data}{a data frame containing the new data for which to generate predictions.}
}
\description{
Define a lightgbm model specification

Fit a lightgbm model based on a specification

Predict lightgbm results on new data
}
\examples{
set.seed(0)
training <- sim_data(setup = "A", n = 300, p = 6, sigma = 1)

trt_mod <- lightgbm_spec(
  formula = training$formulas$treatment,
  mode = "classification",
)

fit_trt <- fit(trt_mod, training$data)

pred_trt_train <- predict(fit_trt, training$data)
# boxplot(pred_trt_train ~ training$data$treatment)

test <- sim_data(setup = "A", n = 2000, p = 6, sigma = 1)
pred_trt_test <- predict(fit_trt, test$data)
# boxplot(pred_trt_test ~ test$data$treatment)


out_mod <- lightgbm_spec(
  formula = training$formulas$outcome,
  mode = "regression",
)
fit_out <- fit(out_mod, training$data)

pred_out_train <- predict(fit_out, training$data)
pred_out_test <- predict(fit_out, test$data)
# plot(pred_out_train, training$data$outcome)
# plot(pred_out_test, test$data$outcome)
}
