% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{multiway_cf_folds}
\alias{multiway_cf_folds}
\title{Higher order cross-fitting instructions with inverted train/test proportions.}
\usage{
multiway_cf_folds(data, v)
}
\arguments{
\item{data}{a data.frame or tibble}

\item{v}{the number of folds to create.}
}
\value{
a rsample rset object with v splits.
}
\description{
Higher order cross-fitting instructions with inverted train/test proportions.
}
