% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{predict.cvboost_fit}
\alias{predict.cvboost_fit}
\title{Predict cvboost results on new data}
\usage{
\method{predict}{cvboost_fit}(object, new_data, ...)
}
\arguments{
\item{object}{A fitted cvboost_fit object.}

\item{new_data}{A data frame containing the new data
on which to make predictions.}

\item{...}{Additional arguments passed to stats::predict.}
}
\description{
Predict cvboost results on new data
}
