% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_learner.R
\name{predict.t_learner}
\alias{predict.t_learner}
\title{Predict the treatment effect using a t-learner}
\usage{
\method{predict}{t_learner}(object, new_data, ...)
}
\arguments{
\item{object}{A fitted t-learner object.}

\item{new_data}{A data frame containing the new data to predict on.}

\item{...}{Additional arguments passed to predict_expected_value.}
}
\description{
Predict the treatment effect using a t-learner
}
