% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{predict_prob}
\alias{predict_prob}
\title{A wrapper for predict that changes handling of default types for workflow}
\usage{
predict_prob(object, ...)
}
\description{
The default \code{type} for prediction for mode = classification
in parsnip is to output predicted classes.
This changes that default to 'prob', which
produces a tibble of probabilities for each class.
}
\details{
For regression modes in tidymodels, or for user defined
model object classes, \code{predict_prob} defaults to \code{predict}.
}
