% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_fun.R
\name{pseudo_DR}
\alias{pseudo_DR}
\alias{pseudo_plugin}
\alias{pseudo_U}
\alias{pseudo_cov}
\title{Pseudo outcome functions}
\usage{
pseudo_DR(
  outcome,
  treatment,
  .pred_outcome_1,
  .pred_outcome_0,
  .pred_treatment,
  .pred_control
)

pseudo_plugin(.pred_outcome_1, .pred_outcome_0)

pseudo_U(outcome, treatment, .pred_outcome_obs, .pred_treatment)

pseudo_cov(outcome, treatment, .pred_outcome_obs, .pred_treatment)
}
\description{
Pseudo outcome functions
}
