% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_functions.R
\name{ptrim}
\alias{ptrim}
\title{Trim values in a numeric vector to be within a specified range}
\usage{
ptrim(x, min_val, max_val)
}
\arguments{
\item{x}{A numeric vector to be trimmed.}

\item{min_val}{The minimum value to which elements of \code{x} should be trimmed.}

\item{max_val}{The maximum value to which elements of \code{x} should be trimmed.}
}
\description{
Trim values in a numeric vector to be within a specified range
}
