% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneflow.R
\name{subsample}
\alias{subsample}
\title{Generate subsamples for tuning}
\usage{
subsample(data, n_analysis, times)
}
\description{
This function is useful for pre-tuning hyperparameters
for samples of different sizes.
}
