% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_learner.R
\name{t_learner}
\alias{t_learner}
\title{Fit a two-model-learner (t-learner)}
\usage{
t_learner(data, outcome_0_wf, outcome_1_wf = outcome_0_wf)
}
\arguments{
\item{data}{A data frame containing the data to fit the t-learner on.}

\item{outcome_0_wf, outcome_1_wf}{Workflows for the outcome models under control and treatment, respectively.}
}
\description{
Fit a two-model-learner (t-learner)
}
