% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneflow.R
\name{tune_wf}
\alias{tune_wf}
\title{Tune a workflow and return the best workflow}
\usage{
tune_wf(wf, data, v = 5, ...)
}
\arguments{
\item{wf}{a parsnip workflow}

\item{data}{a dataset to train on}

\item{v}{number of cv folds to use}

\item{...}{passed to tune_race_anova}
}
\description{
Tune a workflow and return the best workflow
}
