% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{tuned_boost_spec}
\alias{tuned_boost_spec}
\title{Define a boosting specification, tuned by cross-validation}
\usage{
tuned_boost_spec(data = NULL, fit_init = NULL, ...)
}
\arguments{
\item{data}{A data frame containing the data to fit the initial cvboost model on.}

\item{fit_init}{An initial cvboost fit object, if already available.}

\item{...}{Additional arguments passed to cvboost_spec.}
}
\description{
Either data or fit_init must be provided.
}
