% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_params.R
\name{update_params}
\alias{update_params}
\alias{update_params.lightgbm_spec}
\alias{update_params.workflow}
\title{Update the parameters of a model specification or workflow}
\usage{
update_params(object, new_params, ...)

\method{update_params}{lightgbm_spec}(object, new_params, ...)

\method{update_params}{workflow}(object, new_params, ...)
}
\arguments{
\item{object}{A model specification or workflow object to update.}

\item{new_params}{A named list of new parameters to update in the model specification or workflow.}

\item{...}{Placeholder for future arguments, not currently used.}
}
\description{
Update the parameters of a model specification or workflow
}
