% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_fun.R
\name{weight_DR_X}
\alias{weight_DR_X}
\alias{weight_U_X}
\alias{weight_DR_AX}
\alias{weight_DR_alt_AX}
\alias{weight_U_AX}
\alias{weight_1}
\title{Weighting functions}
\usage{
weight_DR_X(.pred_treatment, .pred_control)

weight_U_X(treatment, .pred_treatment, .pred_control)

weight_DR_AX(treatment, .pred_treatment, .pred_control)

weight_DR_alt_AX(treatment, .pred_treatment, .pred_control)

weight_U_AX(treatment, .pred_treatment)

weight_1(.pred_treatment)
}
\arguments{
\item{.pred_treatment, .pred_control}{Vectors of predicted probabilities of treatment and control assignment.}

\item{treatment}{Vector of treatment assignments (1 for treatment, 0 for control).}
}
\description{
Weighting functions
}
