% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{wpor}
\alias{wpor}
\title{Apply weighted pseudo-outcome regression}
\usage{
wpor(
  dat,
  outcome_wf,
  treatment_wf,
  effect_wf,
  pseudo_fun = pseudo_DR,
  weight_fun = weight_1,
  min_prob = 0.01,
  v,
  cf_order = 2,
  verbose = TRUE
)
}
\description{
Apply weighted pseudo-outcome regression
}
