% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_column.R
\name{add_weights_column}
\alias{add_weights_column}
\title{Specify a column name to use for weights}
\usage{
add_weights_column(object, ...)
}
\description{
Specify a column name to use for weights
}
\examples{
training <- sim_data(setup = "A", n = 700, p = 6, sigma = 1)
tdata <- training$data
tdata$w <- 1

gbm <- lightgbm_spec(
  formula = training$formulas$treatment,
  mode = "classification",
)
gbmw <- add_weights_column(gbm, "w")
w <- get_weights(gbmw, tdata)
stopifnot(identical(w, tdata$w))
}
