% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneflow.R
\name{tune_wf}
\alias{tune_wf}
\alias{as.tuneflow}
\alias{fit.tuneflow}
\title{Tune a workflow and return the best workflow}
\usage{
tune_wf(wf, data, v = 5, ...)

as.tuneflow(wf, ...)

\method{fit}{tuneflow}(object, data, ...)
}
\arguments{
\item{wf}{a parsnip workflow}

\item{data}{a dataset to train on}

\item{v}{number of cv folds to use}

\item{...}{passed to fit.workflow (e.g., weights)}
}
\description{
\code{tune_wf}, \code{as.tuneflow} and \code{fit.tuneflow} are convenience functions
for tuning a workflow.
}
\details{
\code{tune_wf} finalizes a workflow using \code{finetune::tune_race_anova},
cross-validation, and either RMSE
or AUC, depending on the workflow's \code{mode}.

\code{as.tuneflow} creates an object of class \code{tuneflow}, which is a list containing
(1) a workflow, and (2) instructions on how to tune it, i.e., arguments to be
passed to tune_wf.

At the time of fitting, \code{fit.tuneflow} first applies tune_wf
to finalize the workflow, using the stored arguments from \code{as.tuneflow},
and then applies \code{fit.workflow} to fit the finalized workflow.
}
\examples{
dontrun{
#' library(tidymodels)
library(dplyr)
data(cells, package = "modeldata")
cells <- cells \%>\% select(-case)

train_data <- sim_data(setup = 'A', n = 300, p = 6, sigma = 1)$data

x_terms <- train_data \%>\%
  select(starts_with("x.")) \%>\%
  colnames
rhs <- paste(x_terms, collapse = " + ")
treatment_formula <- formula(paste("treatment ~", rhs))

mod_spec <- boost_tree(min_n = tune(), learn_rate = tune()) \%>\%
  set_engine("xgboost")

wf <- workflow() \%>\%
  add_model(set_mode(mod_spec, "classification")) \%>\%
  add_formula(treatment_formula)

if(FALSE){
  # Optionally, the tuning process can be parallelized by first
  # registering a cluster.
  cores <- parallel::detectCores(logical = FALSE)
  cl <- makePSOCKcluster(cores)
  registerDoParallel(cl)
}

## Example using tune_wf explicitly
tuned_wf <- tune_wf(wf, data = train_data)
fitted1 <- fit(tuned_wf, train_data)
pred1 <- predict(fitted1, train_data, 'prob')

## Example using tune_wf implicitly
fitted2 <- as.tuneflow(wf) \%>\% 
  wpor::fit(train_data)
pred2 <- predict(fitted1, train_data, 'prob')

range(pred1 - pred2) #equivalent results
}
}
