% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{crossfit_nuisance}
\alias{crossfit_nuisance}
\alias{fit_wpor}
\title{Apply weighted pseudo-outcome regression}
\usage{
crossfit_nuisance(
  data,
  outcome_obs_wf,
  outcome_1_wf,
  outcome_0_wf,
  treatment_wf,
  min_prob = 0.01,
  v = 10,
  cf_order = 2,
  verbose = TRUE
)

fit_wpor(
  data,
  nuisance_tbl = NULL,
  ...,
  pseudo_fun,
  weight_fun,
  effect_wf,
  standardize_weights = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data.frame or tibble containing columns \code{treatment} and \code{outcome}.}

\item{outcome_1_wf, outcome_0_wf, outcome_obs_wf, treatment_wf}{tidmodel workflows for each modeling step. These should have formulas with
\code{outcome} and \code{treatment} on the left-hand side, respectively.}

\item{min_prob}{a truncation probability for the propensity scores.}

\item{v}{number of folds, passed to vfold_cv}

\item{cf_order}{either 2, 3, or 4.}

\item{verbose}{whether to print progress.}

\item{nuisance_tbl}{the output of crossfit_nuisance}

\item{...}{passed to crossfit_nuisance, if nuisance_tbl is not
supplied.}

\item{pseudo_fun}{the pseudo-outcome function. See pseudo_fun.R}

\item{weight_fun}{the weighting function. See weight_fun.R#'}

\item{effect_wf}{A workflow for fitting the final effect (pseudo-outcome) model.}
}
\value{
a tibble of nuisance predictions. If cf_order is >2, there will be more than 1 prediction per row of \code{data}. The tibble contains: \code{.row}, the row index of \code{data}; \code{.fold_id}, the fold used to train the predictions; \code{.pred_treatment}, the predicted probability of treatment; \code{.pred_control}, the predicted probability of control (equal to 1-.pred_treatment if cf_order <= 3); \code{.pred_outcome_0} the predicted outcome under treatment == 0, \code{.pred_outcome_1} the predicted outcome under treatment == 1; \code{.pred_outcome_obs} the predicted outcome under the observed treatment, i.e., marginalizing over treatment.

a fitted model; the output of \code{workflows:::fit.workflow}
}
\description{
Apply weighted pseudo-outcome regression

Fit a weighted pseudo-outcome regression
}
