% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{cvboost_spec}
\alias{cvboost_spec}
\title{Define a cvboost model specification}
\usage{
cvboost_spec(formula, mode, control = NULL, weights_column = NULL)
}
\arguments{
\item{process_x_mat}{a function to be applied to
the result of model.matrix}
}
\description{
Define a cvboost model specification
}
