% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{fit.cvboost_spec}
\alias{fit.cvboost_spec}
\title{Fit a cvboost model based on a specification}
\usage{
\method{fit}{cvboost_spec}(object, data)
}
\description{
Fit a cvboost model based on a specification
}
