% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{fit_on_folds}
\alias{fit_on_folds}
\title{Fit workflow or object on resampled folds}
\usage{
fit_on_folds(object, rset)
}
\arguments{
\item{object}{must have a fit method, which returns an object with a
predict method; which returns a vector of expected values for the predicted outcome.
If the object's predict method returns output in a different format
(e.g., a fitted parsnip model), then the user must define a
\code{predict_expected_value} method.}

\item{rset}{e.g., the output of vfold_cv
This function is based on tune::fit_resamples, but it is not a generic.
Thus, for custom fitting procedures that are not workflows,
the user still needs to ensure that a fit method exists,
but does not need to ensure that a fit_resamples method exists.
This version relies only on the \code{fit} method of the workflow.}
}
\description{
Fit workflow or object on resampled folds
}
