% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{fit_on_folds}
\alias{fit_on_folds}
\title{Based on tune::fit_resamples, but it is not a generic.
Thus, for custom fitting procedures that are not workflows,
the user still needs to ensure that a fit method exists,
but does not need to ensure that a fit_resamples method exists.
This version relies only on the \code{fit} method of the workflow.}
\usage{
fit_on_folds(object, rset)
}
\description{
Based on tune::fit_resamples, but it is not a generic.
Thus, for custom fitting procedures that are not workflows,
the user still needs to ensure that a fit method exists,
but does not need to ensure that a fit_resamples method exists.
This version relies only on the \code{fit} method of the workflow.
}
