% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor.R
\name{multiway_cf_folds}
\alias{multiway_cf_folds}
\title{Higher order cross-fitting instructions with inverted train/test proportions.}
\usage{
multiway_cf_folds(data, v)
}
\description{
Higher order cross-fitting instructions with inverted train/test proportions.
}
