% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{predict.cvboost_fit}
\alias{predict.cvboost_fit}
\title{Predict cvboost results on new data}
\usage{
\method{predict}{cvboost_fit}(cvboost_fit, new_data)
}
\description{
Predict cvboost results on new data
}
