% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_fun.R
\name{pseudo_DR_single}
\alias{pseudo_DR_single}
\alias{pseudo_DR_separate}
\alias{pseudo_plugin}
\alias{pseudo_U}
\alias{pseudo_cov}
\title{Pseudo outcome functions}
\usage{
pseudo_DR_single(
  outcome,
  treatment,
  .pred_outcome_1_single,
  .pred_outcome_0_single,
  .pred_treatment,
  .pred_control
)

pseudo_DR_separate(
  outcome,
  treatment,
  .pred_outcome_1_separate,
  .pred_outcome_0_separate,
  .pred_treatment,
  .pred_control
)

pseudo_plugin(.pred_outcome_1_single, .pred_outcome_0_single)

pseudo_U(outcome, treatment, .pred_outcome_marginal, .pred_treatment)

pseudo_cov(
  outcome,
  treatment,
  .pred_outcome_marginal,
  .pred_treatment,
  .pred_control
)
}
\description{
Pseudo outcome functions
}
