% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost_spec.R
\name{tuned_boost_spec}
\alias{tuned_boost_spec}
\title{Define a boosting specification, tuned by cross-validation}
\usage{
tuned_boost_spec(
  formula,
  mode,
  control = NULL,
  data,
  weights = NULL,
  fit_init = NULL
)
}
\description{
Define a boosting specification, tuned by cross-validation
}
