% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_fun.R
\name{weight_DR_X}
\alias{weight_DR_X}
\alias{weight_U_X}
\alias{weight_DR_AX}
\alias{weight_U_AX}
\alias{weight_1}
\title{Weighting functions}
\usage{
weight_DR_X(.pred_treatment, .pred_control)

weight_U_X(treatment, .pred_treatment, .pred_control)

weight_DR_AX(treatment, .pred_treatment, .pred_control)

weight_U_AX(treatment, .pred_treatment)

weight_1(.pred_treatment)
}
\description{
Weighting functions
}
