% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpor-package.R
\docType{package}
\name{wpor-package}
\alias{wpor}
\alias{wpor-package}
\title{wpor: Weighted pseudo-outcome regression}
\description{
This package implements the weighted pseudo-outcome regression methods discussed in https://arxiv.org/abs/2307.09700. For nuisance training methods, users can either input workflows from the tidymodels set of packages or define their own training algorithms.
}
\keyword{internal}
